Whitespace improvements (#143)
* Add .gitattributes

Make sure that all text files are added with LF as eol to the
repository, regardless of how they appear when checked out.  Currently,
all text files already use only LF (see `git ls-files --eol`), but
`text=auto` prevents potential mishaps if someone ever tries to commit a
file with CRLF.

For detailed explanations, see "End of line conversion" on
gitattributes(5) and the following post:
https://www.hanselman.com/blog/youre-just-another-carriage-return-line-feed-in-the-wall

* Fix whitespace with git stripspace

On all (non-empty) text files (except LICENSE).

Commands used:

    git ls-files --eol | grep -v '^i/-text' |
      cut -f 2 -d "$(printf '\t')" | grep -v LICENSE | tr '\n' '\000' |
      xargs -0 -I '{}' sh -c \
      "test -s '{}' && printf '%s\n' \"\`git stripspace <'{}'\`\" >'{}'"

* .editorconfig: enable trim trailing ws / insert final newline

On all files.

* .editorconfig: define consistent charset and eol

This is already the case on all text files, which can be checked with
the following commands:

    git ls-files --eol | grep -v '^i/-text' |
      cut -f 2 -d "$(printf '\t')" | tr '\n' '\000' | xargs -0 file

So tell editorconfig to help keep it that way.